Fix some types and add logging to DRI
# 50-character subject line
#
# 72-character wrapped longer description. This should answer:
#
# * Why was this change necessary?
# * How does it address the problem?
# * Are there any side effects?
#
# Include a link to the ticket, if any.
#
# Add co-authors if you worked on this code with others:
#

+++ model/Die.c
@@ -16,7 +16,7 @@ Die * die_create(size_t numFaces) {
 }
 
 long long int die_roll(Die *d) {
-	d->last_result = (size_t) arc4random_uniform(d->num_faces) + 1;
+	d->last_result = (long long int) arc4random_uniform(d->num_faces) + 1;
 	return d->last_result;
 }
